day_53/requirements.txt
streamlit==1.45.1
speechrecognition==3.10.0
googletrans==4.0.0-rc1
gTTS==2.5.3
python-dotenv==0.19.0
google-generativeai==0.8.5